Trait system_runtime_primitives::SystemDomainApi

pub trait SystemDomainApi<Block: BlockT, PNumber: Encode + Decode, PHash: Encode + Decode>: Core<Block> {

    // Provided methods
    fn construct_submit_core_bundle_extrinsics(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        signed_opaque_bundles: Vec<SignedOpaqueBundle<PNumber, PHash, <Block as BlockT>::Hash>>
    ) -> Result<Vec<Vec<u8>>, ApiError> { ... }
    fn construct_submit_core_bundle_extrinsics_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        signed_opaque_bundles: Vec<SignedOpaqueBundle<PNumber, PHash, <Block as BlockT>::Hash>>
    ) -> Result<Vec<Vec<u8>>, ApiError> { ... }
    fn bundle_election_solver_params(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId
    ) -> Result<BundleElectionSolverParams, ApiError> { ... }
    fn bundle_election_solver_params_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId
    ) -> Result<BundleElectionSolverParams, ApiError> { ... }
    fn core_bundle_election_storage_keys(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId,
        executor_public_key: ExecutorPublicKey
    ) -> Result<Option<Vec<Vec<u8>>>, ApiError> { ... }
    fn core_bundle_election_storage_keys_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId,
        executor_public_key: ExecutorPublicKey
    ) -> Result<Option<Vec<Vec<u8>>>, ApiError> { ... }
    fn head_receipt_number(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn head_receipt_number_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn oldest_receipt_number_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn maximum_receipt_drift_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<NumberFor<Block>, ApiError> { ... }
    fn submit_fraud_proof_unsigned(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        fraud_proof: FraudProof<NumberFor<Block>, Block::Hash>
    ) -> Result<(), ApiError> { ... }
    fn submit_fraud_proof_unsigned_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        fraud_proof: FraudProof<NumberFor<Block>, Block::Hash>
    ) -> Result<(), ApiError> { ... }
    fn core_domain_state_root_at(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        domain_id: DomainId,
        number: NumberFor<Block>,
        domain_hash: Block::Hash
    ) -> Result<Option<Block::Hash>, ApiError> { ... }
    fn core_domain_state_root_at_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        domain_id: DomainId,
        number: NumberFor<Block>,
        domain_hash: Block::Hash
    ) -> Result<Option<Block::Hash>, ApiError> { ... }

}

API necessary for system domain.

Provided Methods

fn construct_submit_core_bundle_extrinsics( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, signed_opaque_bundles: Vec<SignedOpaqueBundle<PNumber, PHash, <Block as BlockT>::Hash>> ) -> Result<Vec<Vec<u8>>, ApiError>

Wrap the core domain bundles into extrinsics.

fn construct_submit_core_bundle_extrinsics_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, signed_opaque_bundles: Vec<SignedOpaqueBundle<PNumber, PHash, <Block as BlockT>::Hash>> ) -> Result<Vec<Vec<u8>>, ApiError>

Wrap the core domain bundles into extrinsics.

fn bundle_election_solver_params( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId ) -> Result<BundleElectionSolverParams, ApiError>

Returns the parameters for solving the bundle election.

fn bundle_election_solver_params_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId ) -> Result<BundleElectionSolverParams, ApiError>

Returns the parameters for solving the bundle election.

fn core_bundle_election_storage_keys( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId, executor_public_key: ExecutorPublicKey ) -> Result<Option<Vec<Vec<u8>>>, ApiError>

fn core_bundle_election_storage_keys_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId, executor_public_key: ExecutorPublicKey ) -> Result<Option<Vec<Vec<u8>>>, ApiError>

fn head_receipt_number( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

fn head_receipt_number_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

fn oldest_receipt_number( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

fn oldest_receipt_number_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId ) -> Result<NumberFor<Block>, ApiError>

fn maximum_receipt_drift( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<NumberFor<Block>, ApiError>

fn maximum_receipt_drift_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<NumberFor<Block>, ApiError>

fn submit_fraud_proof_unsigned( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, fraud_proof: FraudProof<NumberFor<Block>, Block::Hash> ) -> Result<(), ApiError>

fn submit_fraud_proof_unsigned_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, fraud_proof: FraudProof<NumberFor<Block>, Block::Hash> ) -> Result<(), ApiError>

fn core_domain_state_root_at( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, domain_id: DomainId, number: NumberFor<Block>, domain_hash: Block::Hash ) -> Result<Option<Block::Hash>, ApiError>

Returns the core domain’s state root at a specific number and hash.

fn core_domain_state_root_at_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, domain_id: DomainId, number: NumberFor<Block>, domain_hash: Block::Hash ) -> Result<Option<Block::Hash>, ApiError>

Returns the core domain’s state root at a specific number and hash.